style: project
fix issues pointed out in pull request review
