+++ src/MainPage/MainPage.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Form from './Form.js'
+
+const MainPage = () => {
+    return (
+      <div>
+        <Form></Form>
+      </div>
+    );
+}
+
+export default MainPage;
